Add minimal command to start a notebook execution
`vh notebook run` can be used to start a notebook execution.

This is a minimal working version to be able to start the execution: it
requires 'environment' and 'image' to be defined as command-line
arguments, and otherwise uses default values; there is no customization
of the command.

Adding options to customize the run is planned for later.