Add minimal command to start a notebook execution (#334)
* Add minimal command to start a notebook execution

`vh notebook run` can be used to start a notebook execution.

This is a minimal working version to be able to start the execution: it
requires 'environment' and 'image' to be defined as command-line
arguments, and otherwise uses default values; there is no customization
of the command.

Adding options to customize the run is planned for later.

* Add notebook execution start test

The mock data is from successfully starting a notebook execution via the
API.

The test only includes the minimum required data to start the execution.
More to be added later.